+++ changelog.md
@@ -1,3 +1,15 @@
-V.1 - Menu
+V.1.1 - Menu
 
-V.2 - Map
+V.1.2 - Map
+
+V.2 - Inventory
+added = Empty list for inventory
+added = create two inventories using nested dictionaries, 
+        inventory1 and inventory2
+added = define a function to access the inventory
+added = define a function to add an item to the inventory
+added = define a function to remove an item from the inventory
+removed inventory1 and inventory2
+added = tile for easy accessibility of map
+removed = typos
+removed = settings and description and replaced with tile